% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-nycflights13.r
\name{h2-nycflights13}
\alias{h2-nycflights13}
\alias{nycflights13_h2}
\title{Database versions of the nycflights13 data}
\usage{
nycflights13_h2(use_cache = TRUE, clear_cache = FALSE)
}
\arguments{
\item{use_cache}{Use cached database if present.}

\item{clear_cache}{Remove database from cache if present.}
}
\description{
These functions cache the data from the \code{nycflights13} database in
a local database, for use in examples and vignettes. Indexes are created
to making joining tables on natural keys efficient.
}
\keyword{internal}

