% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_h2.r
\docType{class}
\name{H2JDBCConnection-class}
\alias{H2JDBCConnection-class}
\title{A wrapper class for JDBCConnection objects created in src_h2. Used for H2 specific dispatch.}
\description{
A wrapper class for JDBCConnection objects created in src_h2. Used for H2 specific dispatch.
}

