% Generated by roxygen2 (4.0.2): do not edit by hand
\name{src_h2}
\alias{src_h2}
\alias{src_h2.H2Connection}
\alias{src_h2.character}
\alias{tbl.src_h2}
\title{Connect to a H2 database.}
\usage{
src_h2(x)

\method{src_h2}{character}(uri, user = "sa", password = "", ...)

\method{src_h2}{H2Connection}(con, ...)

\method{tbl}{src_h2}(src, from, ...)
}
\arguments{
\item{uri}{the URI for the database of a path to a file which will be converted into an URI.}

\item{user}{the username}

\item{password}{the password for the user}

\item{con}{a \code{\linkS4class{H2Connection}} object}
}
\description{
Connect to a H2 database.
}

