% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data-nycflights13.r
\name{h2-nycflights13}
\alias{h2-nycflights13}
\alias{nycflights13_h2}
\title{Database versions of the nycflights13 data}
\usage{
nycflights13_h2(path = NULL)
}
\arguments{
\item{path}{location of sqlite database file}
}
\description{
These functions cache the data from the \code{nycflights13} database in
a local database, for use in examples and vignettes. Indexes are created
to making joining tables on natural keys efficient.
}
\keyword{internal}

