% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_h2.r
\name{src_h2}
\alias{src_desc.src_h2}
\alias{src_h2}
\alias{src_h2.H2Connection}
\alias{src_h2.character}
\alias{src_translate_env.src_h2}
\alias{tbl.src_h2}
\title{Create a source object for an H2 database.}
\usage{
src_h2(x, ...)

\method{src_h2}{character}(x, drv_h2 = dbj.h2::driver(), ...)

\method{src_h2}{H2Connection}(x, ...)

\method{tbl}{src_h2}(src, from, ...)

\method{src_desc}{src_h2}(x)

\method{src_translate_env}{src_h2}(x)
}
\arguments{
\item{x}{an URL for the H2 database connection as defined at \url{http://h2database.com/html/features.html#database_url}.
If \code{x} is a path to a local file, than the '.h2.db' suffix, if present, is stripped off automatically.}

\item{...}{for the src, other arguments passed on to the underlying
database connector, \code{dbConnect}. For the tbl, included for
compatibility with the generic, but otherwise ignored.}

\item{drv_h2}{The database driver}

\item{src}{a h2 src created with \code{src_h2}.}

\item{from}{Either a string giving the name of table in database, or
\code{\link{sql}} described a derived table or compound join.}
}
\description{
Create a source object for an H2 database.
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: Create the source from an H2 URL.

\item \code{H2Connection}: Create the source from an JDBC connection.
}}

